backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
astroid==3.0.3
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
Django==4.2.10
django-debug-toolbar==4.3.0
docopt==0.6.2
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2024.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
isort==5.13.2
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
mccabe==0.7.0
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai==1.12.0
# openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==23.2
pillow==10.2.0
pipreqs==0.4.13
pkg==0.2
platformdirs==4.2.0
psycopg2-binary==2.9.6
pydantic==2.6.1
pydantic_core==2.16.2
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
tiktoken==0.6.0
tomli==2.0.1
tomlkit==0.12.3
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
whitenoise==6.5.0
yarg==0.1.9
